avoid implicit conversions

+++ src/sim/interface.cc
@@ -820,7 +820,7 @@ void Interface::execute_run(size_t nb_steps, Glossary& opt)
             evaluate(code, "in run:code");
 
         delta = real(nb_steps) / real(nb_frames);
-        check = (int)delta;
+        check = delta;
     }
     
     simul.prepare();
@@ -841,7 +841,7 @@ void Interface::execute_run(size_t nb_steps, Glossary& opt)
             }
             if ( sss >= nb_steps )
                 break;
-            check = (int)( ++frame * delta );
+            check = ( ++frame * delta );
         }
 
         hold();
